app/requirements.txt
fastapi==0.111.1
fastapi_cache2==0.2.2
pydantic==2.8.2
pytest==8.3.2
python-dotenv==1.0.1
redis==5.0.7
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
alembic==1.13.2
uvicorn==0.30.3
pydantic_settings==2.3.4